// components/user-profile/types.ts
export type EditableProfile = {
  fullName: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  jobTitle: string;
  bio: string;
  avatarUrl: string;
  country: string;
  cityState: string;
  postalCode: string;
  taxId: string;
  facebookUrl: string;
  xUrl: string;
  linkedinUrl: string;
  instagramUrl: string;
};